+++ communications/__init__.py
@@ -1,12 +1,12 @@
 # Copyright (c) 2025, AgriTheory and contributors
 # For license information, please see license.txt
 
-__version__ = "15.2.0"
+__version__ = "15.4.0"
 
 
 # monkey patch to allow extensions in phone number
 # frappe.utils.validate_phone_number = validate_phone_number
-from frappe.desk.form import assign_to
+from frappe.desk.form import assign_to  # nosemgrep: frappe-monkey-patching-not-allowed
 from communications.communications.overrides.assign_to import (
 	custom_notify_assignment,
 )  # noqa: E402